Fix fun. + insert synchronization to Queries.

+++ yea-mobile/src/java/module/Product.java
@@ -126,7 +126,7 @@ public class Product {
 
     public static String edit(int pid) {
 //        Product product = null;
-//        // TODO: return the product from the data base instead of from the static list
+//        // TODO: Assaf - return the product from the data base instead of from the static list
 //        for (Product p : getProductsList()) {
 //            if (pid.equals(p.pid)) {
 //                product = p;
@@ -140,14 +140,18 @@ public class Product {
     }
     
     public static String delete(int pid) {
-        //TODO: @Yishay - if the product is connected to cart, then a bad error is shown
-        // Instead - we need to show a pretty "can't delete because user user has this product in his cart" error
-        Product removedProduct = Queries.getInstance().deleteProduct(pid);
-        if (removedProduct == null) {
-            ErrorReporter.addError("Product doesn't exist in the first place...");
+        Product product= Queries.getInstance().getProduct(pid);
+        if(product!=null){
+            if (product.usersCarts.isEmpty()){
+                Queries.getInstance().deleteProduct(pid);
+                return "/products.xhtml?faces-redirect=true";
+            }
+            ErrorReporter.addError("Can't delete because user user has this product in his cart");
             return "/products.xhtml?faces-redirect=false";
         }
-        return "/products.xhtml?faces-redirect=true";
+            ErrorReporter.addError("Product doesn't exist in the first place...");
+            return "/products.xhtml?faces-redirect=false";
+
     }
